+++ src/main/java/com/boheco1/dev/integratedaccountingsystem/warehouse/UserMyMIRSApplicationController.java
@@ -428,7 +428,7 @@ public class UserMyMIRSApplicationController extends MenuControllerHandler imple
             try {
                 selectedStock = StockDAO.get(result.getId());
                 selectedStock.setQuantity(result.getQuantity());
-                int av = StockDAO.countAvailable(selectedStock);
+                double av = StockDAO.countAvailable(selectedStock);
                 if(av == 0) {
                     AlertDialogBuilder.messgeDialog("System Warning", "Insufficient stock.",
                             Utility.getStackPane(), AlertDialogBuilder.DANGER_DIALOG);
